@@ -1,3 +1,3 @@
 module github.com/bluele/gcache
 
-go 1.15
+go 1.18
