@@ -6,39 +6,37 @@ import (
 )
 
 // Constantly balances between LRU and LFU, to improve the combined result.
-type ARC struct {
-	baseCache
-	items map[interface{}]*arcItem
+type ARC[K comparable, V any, S any] struct {
+	baseCache[K, V, S]
+	items map[K]*arcItem[K, S]
 
 	part int
-	t1   *arcList
-	t2   *arcList
-	b1   *arcList
-	b2   *arcList
+	t1   *arcList[K]
+	t2   *arcList[K]
+	b1   *arcList[K]
+	b2   *arcList[K]
 }
 
-func newARC(cb *CacheBuilder) *ARC {
-	c := &ARC{}
-	buildCache(&c.baseCache, cb)
-
+func newARC[K comparable, V any, S any](cb *CacheBuilder[K, V, S]) *ARC[K, V, S] {
+	c := &ARC[K, V, S]{baseCache: cb.newBaseCache()}
 	c.init()
 	c.loadGroup.cache = c
 	return c
 }
 
-func (c *ARC) init() {
-	c.items = make(map[interface{}]*arcItem)
-	c.t1 = newARCList()
-	c.t2 = newARCList()
-	c.b1 = newARCList()
-	c.b2 = newARCList()
+func (c *ARC[K, V, S]) init() {
+	c.items = make(map[K]*arcItem[K, S])
+	c.t1 = newARCList[K]()
+	c.t2 = newARCList[K]()
+	c.b1 = newARCList[K]()
+	c.b2 = newARCList[K]()
 }
 
-func (c *ARC) replace(key interface{}) {
+func (c *ARC[K, V, S]) replace(key K) {
 	if !c.isCacheFull() {
 		return
 	}
-	var old interface{}
+	var old K
 	if c.t1.Len() > 0 && ((c.b2.Has(key) && c.t1.Len() == c.part) || (c.t1.Len() > c.part)) {
 		old = c.t1.RemoveTail()
 		c.b1.PushFront(old)
@@ -52,13 +50,14 @@ func (c *ARC) replace(key interface{}) {
 	item, ok := c.items[old]
 	if ok {
 		delete(c.items, old)
-		if c.evictedFunc != nil {
-			c.evictedFunc(item.key, item.value)
+		if err := c.evictValue(item.key, item.value); err != nil {
+			// TODO: log error
+			_ = err
 		}
 	}
 }
 
-func (c *ARC) Set(key, value interface{}) error {
+func (c *ARC[K, V, S]) Set(key K, value V) error {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 	_, err := c.set(key, value)
@@ -66,7 +65,7 @@ func (c *ARC) Set(key, value interface{}) error {
 }
 
 // Set a new key-value pair with an expiration time
-func (c *ARC) SetWithExpire(key, value interface{}, expiration time.Duration) error {
+func (c *ARC[K, V, S]) SetWithExpire(key K, value V, expiration time.Duration) error {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 	item, err := c.set(key, value)
@@ -75,27 +74,24 @@ func (c *ARC) SetWithExpire(key, value interface{}, expiration time.Duration) er
 	}
 
 	t := c.clock.Now().Add(expiration)
-	item.(*arcItem).expiration = &t
+	item.expiration = &t
 	return nil
 }
 
-func (c *ARC) set(key, value interface{}) (interface{}, error) {
-	var err error
-	if c.serializeFunc != nil {
-		value, err = c.serializeFunc(key, value)
-		if err != nil {
-			return nil, err
-		}
+func (c *ARC[K, V, S]) set(key K, value V) (*arcItem[K, S], error) {
+	serializedValue, err := c.serializeValue(key, value)
+	if err != nil {
+		return nil, err
 	}
 
 	item, ok := c.items[key]
 	if ok {
-		item.value = value
+		item.value = serializedValue
 	} else {
-		item = &arcItem{
+		item = &arcItem[K, S]{
 			clock: c.clock,
 			key:   key,
-			value: value,
+			value: serializedValue,
 		}
 		c.items[key] = item
 	}
@@ -140,8 +136,8 @@ func (c *ARC) set(key, value interface{}) (interface{}, error) {
 			item, ok := c.items[pop]
 			if ok {
 				delete(c.items, pop)
-				if c.evictedFunc != nil {
-					c.evictedFunc(item.key, item.value)
+				if err := c.evictValue(item.key, item.value); err != nil {
+					return nil, err
 				}
 			}
 		}
@@ -163,7 +159,7 @@ func (c *ARC) set(key, value interface{}) (interface{}, error) {
 }
 
 // Get a value from cache pool using key if it exists. If not exists and it has LoaderFunc, it will generate the value using you have specified LoaderFunc method returns value.
-func (c *ARC) Get(key interface{}) (interface{}, error) {
+func (c *ARC[K, V, S]) Get(key K) (V, error) {
 	v, err := c.get(key, false)
 	if err == KeyNotFoundError {
 		return c.getWithLoader(key, true)
@@ -174,7 +170,7 @@ func (c *ARC) Get(key interface{}) (interface{}, error) {
 // GetIFPresent gets a value from cache pool using key if it exists.
 // If it does not exists key, returns KeyNotFoundError.
 // And send a request which refresh value for specified key if cache object has LoaderFunc.
-func (c *ARC) GetIFPresent(key interface{}) (interface{}, error) {
+func (c *ARC[K, V, S]) GetIFPresent(key K) (V, error) {
 	v, err := c.get(key, false)
 	if err == KeyNotFoundError {
 		return c.getWithLoader(key, false)
@@ -182,18 +178,16 @@ func (c *ARC) GetIFPresent(key interface{}) (interface{}, error) {
 	return v, err
 }
 
-func (c *ARC) get(key interface{}, onLoad bool) (interface{}, error) {
+func (c *ARC[K, V, S]) get(key K, onLoad bool) (V, error) {
 	v, err := c.getValue(key, onLoad)
 	if err != nil {
-		return nil, err
+		var zero V
+		return zero, err
 	}
-	if c.deserializeFunc != nil {
-		return c.deserializeFunc(key, v)
-	}
-	return v, nil
+	return c.deserializeValue(key, v)
 }
 
-func (c *ARC) getValue(key interface{}, onLoad bool) (interface{}, error) {
+func (c *ARC[K, V, S]) getValue(key K, onLoad bool) (S, error) {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 	if elt := c.t1.Lookup(key); elt != nil {
@@ -208,8 +202,9 @@ func (c *ARC) getValue(key interface{}, onLoad bool) (interface{}, error) {
 		} else {
 			delete(c.items, key)
 			c.b1.PushFront(key)
-			if c.evictedFunc != nil {
-				c.evictedFunc(item.key, item.value)
+			if err := c.evictValue(item.key, item.value); err != nil {
+				var zero S
+				return zero, err
 			}
 		}
 	}
@@ -225,8 +220,9 @@ func (c *ARC) getValue(key interface{}, onLoad bool) (interface{}, error) {
 			delete(c.items, key)
 			c.t2.Remove(key, elt)
 			c.b2.PushFront(key)
-			if c.evictedFunc != nil {
-				c.evictedFunc(item.key, item.value)
+			if err := c.evictValue(item.key, item.value); err != nil {
+				var zero S
+				return zero, err
 			}
 		}
 	}
@@ -234,44 +230,49 @@ func (c *ARC) getValue(key interface{}, onLoad bool) (interface{}, error) {
 	if !onLoad {
 		c.stats.IncrMissCount()
 	}
-	return nil, KeyNotFoundError
+	var zero S
+	return zero, KeyNotFoundError
 }
 
-func (c *ARC) getWithLoader(key interface{}, isWait bool) (interface{}, error) {
+func (c *ARC[K, V, S]) getWithLoader(key K, isWait bool) (V, error) {
 	if c.loaderExpireFunc == nil {
-		return nil, KeyNotFoundError
+		var zero V
+		return zero, KeyNotFoundError
 	}
-	value, _, err := c.load(key, func(v interface{}, expiration *time.Duration, e error) (interface{}, error) {
+	value, _, err := c.load(key, func(v V, expiration *time.Duration, e error) (V, error) {
 		if e != nil {
-			return nil, e
+			var zero V
+			return zero, e
 		}
 		c.mu.Lock()
 		defer c.mu.Unlock()
 		item, err := c.set(key, v)
 		if err != nil {
-			return nil, err
+			var zero V
+			return zero, err
 		}
 		if expiration != nil {
 			t := c.clock.Now().Add(*expiration)
-			item.(*arcItem).expiration = &t
+			item.expiration = &t
 		}
 		return v, nil
 	}, isWait)
 	if err != nil {
-		return nil, err
+		var zero V
+		return zero, err
 	}
 	return value, nil
 }
 
 // Has checks if key exists in cache
-func (c *ARC) Has(key interface{}) bool {
+func (c *ARC[K, V, S]) Has(key K) bool {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
 	now := time.Now()
 	return c.has(key, &now)
 }
 
-func (c *ARC) has(key interface{}, now *time.Time) bool {
+func (c *ARC[K, V, S]) has(key K, now *time.Time) bool {
 	item, ok := c.items[key]
 	if !ok {
 		return false
@@ -280,21 +281,22 @@ func (c *ARC) has(key interface{}, now *time.Time) bool {
 }
 
 // Remove removes the provided key from the cache.
-func (c *ARC) Remove(key interface{}) bool {
+func (c *ARC[K, V, S]) Remove(key K) bool {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 
 	return c.remove(key)
 }
 
-func (c *ARC) remove(key interface{}) bool {
+func (c *ARC[K, V, S]) remove(key K) bool {
 	if elt := c.t1.Lookup(key); elt != nil {
 		c.t1.Remove(key, elt)
 		item := c.items[key]
 		delete(c.items, key)
 		c.b1.PushFront(key)
-		if c.evictedFunc != nil {
-			c.evictedFunc(key, item.value)
+		if err := c.evictValue(key, item.value); err != nil {
+			// TODO: log error
+			_ = err
 		}
 		return true
 	}
@@ -304,8 +306,9 @@ func (c *ARC) remove(key interface{}) bool {
 		item := c.items[key]
 		delete(c.items, key)
 		c.b2.PushFront(key)
-		if c.evictedFunc != nil {
-			c.evictedFunc(key, item.value)
+		if err := c.evictValue(key, item.value); err != nil {
+			// TODO: log error
+			_ = err
 		}
 		return true
 	}
@@ -314,24 +317,28 @@ func (c *ARC) remove(key interface{}) bool {
 }
 
 // GetALL returns all key-value pairs in the cache.
-func (c *ARC) GetALL(checkExpired bool) map[interface{}]interface{} {
+func (c *ARC[K, V, S]) GetALL(checkExpired bool) map[K]V {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
-	items := make(map[interface{}]interface{}, len(c.items))
+	items := make(map[K]V, len(c.items))
 	now := time.Now()
 	for k, item := range c.items {
 		if !checkExpired || c.has(k, &now) {
-			items[k] = item.value
+			value, err := c.deserializeValue(k, item.value)
+			if err != nil {
+				continue
+			}
+			items[k] = value
 		}
 	}
 	return items
 }
 
 // Keys returns a slice of the keys in the cache.
-func (c *ARC) Keys(checkExpired bool) []interface{} {
+func (c *ARC[K, V, S]) Keys(checkExpired bool) []K {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
-	keys := make([]interface{}, 0, len(c.items))
+	keys := make([]K, 0, len(c.items))
 	now := time.Now()
 	for k := range c.items {
 		if !checkExpired || c.has(k, &now) {
@@ -342,7 +349,7 @@ func (c *ARC) Keys(checkExpired bool) []interface{} {
 }
 
 // Len returns the number of items in the cache.
-func (c *ARC) Len(checkExpired bool) int {
+func (c *ARC[K, V, S]) Len(checkExpired bool) int {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
 	if !checkExpired {
@@ -359,31 +366,30 @@ func (c *ARC) Len(checkExpired bool) int {
 }
 
 // Purge is used to completely clear the cache
-func (c *ARC) Purge() {
+func (c *ARC[K, V, S]) Purge() {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 
-	if c.purgeVisitorFunc != nil {
-		for _, item := range c.items {
-			c.purgeVisitorFunc(item.key, item.value)
-		}
+	for key, item := range c.items {
+		err := c.purgeValue(key, item.value)
+		_ = err // TODO: log error
 	}
 
 	c.init()
 }
 
-func (c *ARC) setPart(p int) {
+func (c *ARC[K, V, S]) setPart(p int) {
 	if c.isCacheFull() {
 		c.part = p
 	}
 }
 
-func (c *ARC) isCacheFull() bool {
+func (c *ARC[K, V, S]) isCacheFull() bool {
 	return (c.t1.Len() + c.t2.Len()) == c.size
 }
 
 // IsExpired returns boolean value whether this item is expired or not.
-func (it *arcItem) IsExpired(now *time.Time) bool {
+func (it *arcItem[K, S]) IsExpired(now *time.Time) bool {
 	if it.expiration == nil {
 		return false
 	}
@@ -394,40 +400,40 @@ func (it *arcItem) IsExpired(now *time.Time) bool {
 	return it.expiration.Before(*now)
 }
 
-type arcList struct {
+type arcList[K comparable] struct {
 	l    *list.List
-	keys map[interface{}]*list.Element
+	keys map[K]*list.Element
 }
 
-type arcItem struct {
+type arcItem[K comparable, S any] struct {
 	clock      Clock
-	key        interface{}
-	value      interface{}
+	key        K
+	value      S
 	expiration *time.Time
 }
 
-func newARCList() *arcList {
-	return &arcList{
+func newARCList[K comparable]() *arcList[K] {
+	return &arcList[K]{
 		l:    list.New(),
-		keys: make(map[interface{}]*list.Element),
+		keys: make(map[K]*list.Element),
 	}
 }
 
-func (al *arcList) Has(key interface{}) bool {
+func (al *arcList[K]) Has(key K) bool {
 	_, ok := al.keys[key]
 	return ok
 }
 
-func (al *arcList) Lookup(key interface{}) *list.Element {
+func (al *arcList[K]) Lookup(key K) *list.Element {
 	elt := al.keys[key]
 	return elt
 }
 
-func (al *arcList) MoveToFront(elt *list.Element) {
+func (al *arcList[K]) MoveToFront(elt *list.Element) {
 	al.l.MoveToFront(elt)
 }
 
-func (al *arcList) PushFront(key interface{}) {
+func (al *arcList[K]) PushFront(key K) {
 	if elt, ok := al.keys[key]; ok {
 		al.l.MoveToFront(elt)
 		return
@@ -436,21 +442,21 @@ func (al *arcList) PushFront(key interface{}) {
 	al.keys[key] = elt
 }
 
-func (al *arcList) Remove(key interface{}, elt *list.Element) {
+func (al *arcList[K]) Remove(key K, elt *list.Element) {
 	delete(al.keys, key)
 	al.l.Remove(elt)
 }
 
-func (al *arcList) RemoveTail() interface{} {
+func (al *arcList[K]) RemoveTail() K {
 	elt := al.l.Back()
 	al.l.Remove(elt)
 
-	key := elt.Value
+	key := elt.Value.(K)
 	delete(al.keys, key)
 
 	return key
 }
 
-func (al *arcList) Len() int {
+func (al *arcList[K]) Len() int {
 	return al.l.Len()
 }
