@@ -6,50 +6,45 @@ import (
 )
 
 // Discards the least recently used items first.
-type LRUCache struct {
-	baseCache
-	items     map[interface{}]*list.Element
+type LRUCache[K comparable, V any, S any] struct {
+	baseCache[K, V, S]
+	items     map[K]*list.Element
 	evictList *list.List
 }
 
-func newLRUCache(cb *CacheBuilder) *LRUCache {
-	c := &LRUCache{}
-	buildCache(&c.baseCache, cb)
-
+func newLRUCache[K comparable, V any, S any](cb *CacheBuilder[K, V, S]) *LRUCache[K, V, S] {
+	c := &LRUCache[K, V, S]{baseCache: cb.newBaseCache()}
 	c.init()
 	c.loadGroup.cache = c
 	return c
 }
 
-func (c *LRUCache) init() {
+func (c *LRUCache[K, V, S]) init() {
 	c.evictList = list.New()
-	c.items = make(map[interface{}]*list.Element, c.size+1)
+	c.items = make(map[K]*list.Element, c.size+1)
 }
 
-func (c *LRUCache) set(key, value interface{}) (interface{}, error) {
-	var err error
-	if c.serializeFunc != nil {
-		value, err = c.serializeFunc(key, value)
-		if err != nil {
-			return nil, err
-		}
+func (c *LRUCache[K, V, S]) set(key K, value V) (*lruItem[K, S], error) {
+	serializedValue, err := c.serializeValue(key, value)
+	if err != nil {
+		return nil, err
 	}
 
 	// Check for existing item
-	var item *lruItem
+	var item *lruItem[K, S]
 	if it, ok := c.items[key]; ok {
 		c.evictList.MoveToFront(it)
-		item = it.Value.(*lruItem)
-		item.value = value
+		item = it.Value.(*lruItem[K, S])
+		item.value = serializedValue
 	} else {
 		// Verify size not exceeded
 		if c.evictList.Len() >= c.size {
 			c.evict(1)
 		}
-		item = &lruItem{
+		item = &lruItem[K, S]{
 			clock: c.clock,
 			key:   key,
-			value: value,
+			value: serializedValue,
 		}
 		c.items[key] = c.evictList.PushFront(item)
 	}
@@ -67,7 +62,7 @@ func (c *LRUCache) set(key, value interface{}) (interface{}, error) {
 }
 
 // set a new key-value pair
-func (c *LRUCache) Set(key, value interface{}) error {
+func (c *LRUCache[K, V, S]) Set(key K, value V) error {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 	_, err := c.set(key, value)
@@ -75,7 +70,7 @@ func (c *LRUCache) Set(key, value interface{}) error {
 }
 
 // Set a new key-value pair with an expiration time
-func (c *LRUCache) SetWithExpire(key, value interface{}, expiration time.Duration) error {
+func (c *LRUCache[K, V, S]) SetWithExpire(key K, value V, expiration time.Duration) error {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 	item, err := c.set(key, value)
@@ -84,14 +79,14 @@ func (c *LRUCache) SetWithExpire(key, value interface{}, expiration time.Duratio
 	}
 
 	t := c.clock.Now().Add(expiration)
-	item.(*lruItem).expiration = &t
+	item.expiration = &t
 	return nil
 }
 
 // Get a value from cache pool using key if it exists.
 // If it does not exists key and has LoaderFunc,
 // generate a value using `LoaderFunc` method returns value.
-func (c *LRUCache) Get(key interface{}) (interface{}, error) {
+func (c *LRUCache[K, V, S]) Get(key K) (V, error) {
 	v, err := c.get(key, false)
 	if err == KeyNotFoundError {
 		return c.getWithLoader(key, true)
@@ -102,7 +97,7 @@ func (c *LRUCache) Get(key interface{}) (interface{}, error) {
 // GetIFPresent gets a value from cache pool using key if it exists.
 // If it does not exists key, returns KeyNotFoundError.
 // And send a request which refresh value for specified key if cache object has LoaderFunc.
-func (c *LRUCache) GetIFPresent(key interface{}) (interface{}, error) {
+func (c *LRUCache[K, V, S]) GetIFPresent(key K) (V, error) {
 	v, err := c.get(key, false)
 	if err == KeyNotFoundError {
 		return c.getWithLoader(key, false)
@@ -110,22 +105,20 @@ func (c *LRUCache) GetIFPresent(key interface{}) (interface{}, error) {
 	return v, err
 }
 
-func (c *LRUCache) get(key interface{}, onLoad bool) (interface{}, error) {
+func (c *LRUCache[K, V, S]) get(key K, onLoad bool) (V, error) {
 	v, err := c.getValue(key, onLoad)
 	if err != nil {
-		return nil, err
+		var zero V
+		return zero, err
 	}
-	if c.deserializeFunc != nil {
-		return c.deserializeFunc(key, v)
-	}
-	return v, nil
+	return c.deserializeValue(key, v)
 }
 
-func (c *LRUCache) getValue(key interface{}, onLoad bool) (interface{}, error) {
+func (c *LRUCache[K, V, S]) getValue(key K, onLoad bool) (S, error) {
 	c.mu.Lock()
 	item, ok := c.items[key]
 	if ok {
-		it := item.Value.(*lruItem)
+		it := item.Value.(*lruItem[K, S])
 		if !it.IsExpired(nil) {
 			c.evictList.MoveToFront(item)
 			v := it.value
@@ -141,37 +134,42 @@ func (c *LRUCache) getValue(key interface{}, onLoad bool) (interface{}, error) {
 	if !onLoad {
 		c.stats.IncrMissCount()
 	}
-	return nil, KeyNotFoundError
+	var zero S
+	return zero, KeyNotFoundError
 }
 
-func (c *LRUCache) getWithLoader(key interface{}, isWait bool) (interface{}, error) {
+func (c *LRUCache[K, V, S]) getWithLoader(key K, isWait bool) (V, error) {
 	if c.loaderExpireFunc == nil {
-		return nil, KeyNotFoundError
+		var zero V
+		return zero, KeyNotFoundError
 	}
-	value, _, err := c.load(key, func(v interface{}, expiration *time.Duration, e error) (interface{}, error) {
+	value, _, err := c.load(key, func(v V, expiration *time.Duration, e error) (V, error) {
 		if e != nil {
-			return nil, e
+			var zero V
+			return zero, e
 		}
 		c.mu.Lock()
 		defer c.mu.Unlock()
 		item, err := c.set(key, v)
 		if err != nil {
-			return nil, err
+			var zero V
+			return zero, err
 		}
 		if expiration != nil {
 			t := c.clock.Now().Add(*expiration)
-			item.(*lruItem).expiration = &t
+			item.expiration = &t
 		}
 		return v, nil
 	}, isWait)
 	if err != nil {
-		return nil, err
+		var zero V
+		return zero, err
 	}
 	return value, nil
 }
 
 // evict removes the oldest item from the cache.
-func (c *LRUCache) evict(count int) {
+func (c *LRUCache[K, V, S]) evict(count int) {
 	for i := 0; i < count; i++ {
 		ent := c.evictList.Back()
 		if ent == nil {
@@ -183,30 +181,30 @@ func (c *LRUCache) evict(count int) {
 }
 
 // Has checks if key exists in cache
-func (c *LRUCache) Has(key interface{}) bool {
+func (c *LRUCache[K, V, S]) Has(key K) bool {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
 	now := time.Now()
 	return c.has(key, &now)
 }
 
-func (c *LRUCache) has(key interface{}, now *time.Time) bool {
+func (c *LRUCache[K, V, S]) has(key K, now *time.Time) bool {
 	item, ok := c.items[key]
 	if !ok {
 		return false
 	}
-	return !item.Value.(*lruItem).IsExpired(now)
+	return !item.Value.(*lruItem[K, S]).IsExpired(now)
 }
 
 // Remove removes the provided key from the cache.
-func (c *LRUCache) Remove(key interface{}) bool {
+func (c *LRUCache[K, V, S]) Remove(key K) bool {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 
 	return c.remove(key)
 }
 
-func (c *LRUCache) remove(key interface{}) bool {
+func (c *LRUCache[K, V, S]) remove(key K) bool {
 	if ent, ok := c.items[key]; ok {
 		c.removeElement(ent)
 		return true
@@ -214,20 +212,20 @@ func (c *LRUCache) remove(key interface{}) bool {
 	return false
 }
 
-func (c *LRUCache) removeElement(e *list.Element) {
+func (c *LRUCache[K, V, S]) removeElement(e *list.Element) {
 	c.evictList.Remove(e)
-	entry := e.Value.(*lruItem)
+	entry := e.Value.(*lruItem[K, S])
 	delete(c.items, entry.key)
-	if c.evictedFunc != nil {
-		entry := e.Value.(*lruItem)
-		c.evictedFunc(entry.key, entry.value)
+	if err := c.evictValue(entry.key, entry.value); err != nil {
+		// TODO: log error
+		_ = err
 	}
 }
 
-func (c *LRUCache) keys() []interface{} {
+func (c *LRUCache[K, V, S]) keys() []K {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
-	keys := make([]interface{}, len(c.items))
+	keys := make([]K, len(c.items))
 	var i = 0
 	for k := range c.items {
 		keys[i] = k
@@ -237,24 +235,29 @@ func (c *LRUCache) keys() []interface{} {
 }
 
 // GetALL returns all key-value pairs in the cache.
-func (c *LRUCache) GetALL(checkExpired bool) map[interface{}]interface{} {
+func (c *LRUCache[K, V, S]) GetALL(checkExpired bool) map[K]V {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
-	items := make(map[interface{}]interface{}, len(c.items))
+	items := make(map[K]V, len(c.items))
 	now := time.Now()
 	for k, item := range c.items {
 		if !checkExpired || c.has(k, &now) {
-			items[k] = item.Value.(*lruItem).value
+			value, err := c.deserializeValue(k, item.Value.(*lruItem[K, S]).value)
+			if err != nil {
+				// TODO: log error
+				continue
+			}
+			items[k] = value
 		}
 	}
 	return items
 }
 
 // Keys returns a slice of the keys in the cache.
-func (c *LRUCache) Keys(checkExpired bool) []interface{} {
+func (c *LRUCache[K, V, S]) Keys(checkExpired bool) []K {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
-	keys := make([]interface{}, 0, len(c.items))
+	keys := make([]K, 0, len(c.items))
 	now := time.Now()
 	for k := range c.items {
 		if !checkExpired || c.has(k, &now) {
@@ -265,7 +268,7 @@ func (c *LRUCache) Keys(checkExpired bool) []interface{} {
 }
 
 // Len returns the number of items in the cache.
-func (c *LRUCache) Len(checkExpired bool) int {
+func (c *LRUCache[K, V, S]) Len(checkExpired bool) int {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
 	if !checkExpired {
@@ -282,30 +285,29 @@ func (c *LRUCache) Len(checkExpired bool) int {
 }
 
 // Completely clear the cache
-func (c *LRUCache) Purge() {
+func (c *LRUCache[K, V, S]) Purge() {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 
-	if c.purgeVisitorFunc != nil {
-		for key, item := range c.items {
-			it := item.Value.(*lruItem)
-			v := it.value
-			c.purgeVisitorFunc(key, v)
-		}
+	for key, item := range c.items {
+		it := item.Value.(*lruItem[K, S])
+		v := it.value
+		err := c.purgeValue(key, v)
+		_ = err // TODO: log error
 	}
 
 	c.init()
 }
 
-type lruItem struct {
+type lruItem[K comparable, S any] struct {
 	clock      Clock
-	key        interface{}
-	value      interface{}
+	key        K
+	value      S
 	expiration *time.Time
 }
 
 // IsExpired returns boolean value whether this item is expired or not.
-func (it *lruItem) IsExpired(now *time.Time) bool {
+func (it *lruItem[K, S]) IsExpired(now *time.Time) bool {
 	if it.expiration == nil {
 		return false
 	}
