@@ -5,30 +5,28 @@ import (
 )
 
 // SimpleCache has no clear priority for evict cache. It depends on key-value map order.
-type SimpleCache struct {
-	baseCache
-	items map[interface{}]*simpleItem
+type SimpleCache[K comparable, V any, S any] struct {
+	baseCache[K, V, S]
+	items map[K]*simpleItem[S]
 }
 
-func newSimpleCache(cb *CacheBuilder) *SimpleCache {
-	c := &SimpleCache{}
-	buildCache(&c.baseCache, cb)
-
+func newSimpleCache[K comparable, V any, S any](cb *CacheBuilder[K, V, S]) *SimpleCache[K, V, S] {
+	c := &SimpleCache[K, V, S]{baseCache: cb.newBaseCache()}
 	c.init()
 	c.loadGroup.cache = c
 	return c
 }
 
-func (c *SimpleCache) init() {
+func (c *SimpleCache[K, V, S]) init() {
 	if c.size <= 0 {
-		c.items = make(map[interface{}]*simpleItem)
+		c.items = make(map[K]*simpleItem[S])
 	} else {
-		c.items = make(map[interface{}]*simpleItem, c.size)
+		c.items = make(map[K]*simpleItem[S], c.size)
 	}
 }
 
 // Set a new key-value pair
-func (c *SimpleCache) Set(key, value interface{}) error {
+func (c *SimpleCache[K, V, S]) Set(key K, value V) error {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 	_, err := c.set(key, value)
@@ -36,7 +34,7 @@ func (c *SimpleCache) Set(key, value interface{}) error {
 }
 
 // Set a new key-value pair with an expiration time
-func (c *SimpleCache) SetWithExpire(key, value interface{}, expiration time.Duration) error {
+func (c *SimpleCache[K, V, S]) SetWithExpire(key K, value V, expiration time.Duration) error {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 	item, err := c.set(key, value)
@@ -45,31 +43,28 @@ func (c *SimpleCache) SetWithExpire(key, value interface{}, expiration time.Dura
 	}
 
 	t := c.clock.Now().Add(expiration)
-	item.(*simpleItem).expiration = &t
+	item.(*simpleItem[V]).expiration = &t
 	return nil
 }
 
-func (c *SimpleCache) set(key, value interface{}) (interface{}, error) {
-	var err error
-	if c.serializeFunc != nil {
-		value, err = c.serializeFunc(key, value)
-		if err != nil {
-			return nil, err
-		}
+func (c *SimpleCache[K, V, S]) set(key K, value V) (any, error) {
+	serializedValue, err := c.serializeValue(key, value)
+	if err != nil {
+		return nil, err
 	}
 
 	// Check for existing item
 	item, ok := c.items[key]
 	if ok {
-		item.value = value
+		item.value = serializedValue
 	} else {
 		// Verify size not exceeded
 		if (len(c.items) >= c.size) && c.size > 0 {
 			c.evict(1)
 		}
-		item = &simpleItem{
+		item = &simpleItem[S]{
 			clock: c.clock,
-			value: value,
+			value: serializedValue,
 		}
 		c.items[key] = item
 	}
@@ -89,7 +84,7 @@ func (c *SimpleCache) set(key, value interface{}) (interface{}, error) {
 // Get a value from cache pool using key if it exists.
 // If it does not exists key and has LoaderFunc,
 // generate a value using `LoaderFunc` method returns value.
-func (c *SimpleCache) Get(key interface{}) (interface{}, error) {
+func (c *SimpleCache[K, V, S]) Get(key K) (V, error) {
 	v, err := c.get(key, false)
 	if err == KeyNotFoundError {
 		return c.getWithLoader(key, true)
@@ -100,7 +95,7 @@ func (c *SimpleCache) Get(key interface{}) (interface{}, error) {
 // GetIFPresent gets a value from cache pool using key if it exists.
 // If it does not exists key, returns KeyNotFoundError.
 // And send a request which refresh value for specified key if cache object has LoaderFunc.
-func (c *SimpleCache) GetIFPresent(key interface{}) (interface{}, error) {
+func (c *SimpleCache[K, V, S]) GetIFPresent(key K) (V, error) {
 	v, err := c.get(key, false)
 	if err == KeyNotFoundError {
 		return c.getWithLoader(key, false)
@@ -108,18 +103,16 @@ func (c *SimpleCache) GetIFPresent(key interface{}) (interface{}, error) {
 	return v, nil
 }
 
-func (c *SimpleCache) get(key interface{}, onLoad bool) (interface{}, error) {
+func (c *SimpleCache[K, V, S]) get(key K, onLoad bool) (V, error) {
 	v, err := c.getValue(key, onLoad)
 	if err != nil {
-		return nil, err
+		var zero V
+		return zero, err
 	}
-	if c.deserializeFunc != nil {
-		return c.deserializeFunc(key, v)
-	}
-	return v, nil
+	return c.deserializeValue(key, v)
 }
 
-func (c *SimpleCache) getValue(key interface{}, onLoad bool) (interface{}, error) {
+func (c *SimpleCache[K, V, S]) getValue(key K, onLoad bool) (S, error) {
 	c.mu.Lock()
 	item, ok := c.items[key]
 	if ok {
@@ -137,36 +130,41 @@ func (c *SimpleCache) getValue(key interface{}, onLoad bool) (interface{}, error
 	if !onLoad {
 		c.stats.IncrMissCount()
 	}
-	return nil, KeyNotFoundError
+	var zero S
+	return zero, KeyNotFoundError
 }
 
-func (c *SimpleCache) getWithLoader(key interface{}, isWait bool) (interface{}, error) {
+func (c *SimpleCache[K, V, S]) getWithLoader(key K, isWait bool) (V, error) {
 	if c.loaderExpireFunc == nil {
-		return nil, KeyNotFoundError
+		var zero V
+		return zero, KeyNotFoundError
 	}
-	value, _, err := c.load(key, func(v interface{}, expiration *time.Duration, e error) (interface{}, error) {
+	value, _, err := c.load(key, func(v V, expiration *time.Duration, e error) (V, error) {
 		if e != nil {
-			return nil, e
+			var zero V
+			return zero, e
 		}
 		c.mu.Lock()
 		defer c.mu.Unlock()
 		item, err := c.set(key, v)
 		if err != nil {
-			return nil, err
+			var zero V
+			return zero, err
 		}
 		if expiration != nil {
 			t := c.clock.Now().Add(*expiration)
-			item.(*simpleItem).expiration = &t
+			item.(*simpleItem[V]).expiration = &t
 		}
 		return v, nil
 	}, isWait)
 	if err != nil {
-		return nil, err
+		var zero V
+		return zero, err
 	}
 	return value, nil
 }
 
-func (c *SimpleCache) evict(count int) {
+func (c *SimpleCache[K, V, S]) evict(count int) {
 	now := c.clock.Now()
 	current := 0
 	for key, item := range c.items {
@@ -181,14 +179,14 @@ func (c *SimpleCache) evict(count int) {
 }
 
 // Has checks if key exists in cache
-func (c *SimpleCache) Has(key interface{}) bool {
+func (c *SimpleCache[K, V, S]) Has(key K) bool {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
 	now := time.Now()
 	return c.has(key, &now)
 }
 
-func (c *SimpleCache) has(key interface{}, now *time.Time) bool {
+func (c *SimpleCache[K, V, S]) has(key K, now *time.Time) bool {
 	item, ok := c.items[key]
 	if !ok {
 		return false
@@ -197,19 +195,21 @@ func (c *SimpleCache) has(key interface{}, now *time.Time) bool {
 }
 
 // Remove removes the provided key from the cache.
-func (c *SimpleCache) Remove(key interface{}) bool {
+func (c *SimpleCache[K, V, S]) Remove(key K) bool {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 
 	return c.remove(key)
 }
 
-func (c *SimpleCache) remove(key interface{}) bool {
+func (c *SimpleCache[K, V, S]) remove(key K) bool {
 	item, ok := c.items[key]
 	if ok {
 		delete(c.items, key)
-		if c.evictedFunc != nil {
-			c.evictedFunc(key, item.value)
+
+		if err := c.evictValue(key, item.value); err != nil {
+			// TODO: log error
+			_ = err
 		}
 		return true
 	}
@@ -217,10 +217,10 @@ func (c *SimpleCache) remove(key interface{}) bool {
 }
 
 // Returns a slice of the keys in the cache.
-func (c *SimpleCache) keys() []interface{} {
+func (c *SimpleCache[K, V, S]) keys() []K {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
-	keys := make([]interface{}, len(c.items))
+	keys := make([]K, len(c.items))
 	var i = 0
 	for k := range c.items {
 		keys[i] = k
@@ -230,24 +230,29 @@ func (c *SimpleCache) keys() []interface{} {
 }
 
 // GetALL returns all key-value pairs in the cache.
-func (c *SimpleCache) GetALL(checkExpired bool) map[interface{}]interface{} {
+func (c *SimpleCache[K, V, S]) GetALL(checkExpired bool) map[K]V {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
-	items := make(map[interface{}]interface{}, len(c.items))
+	items := make(map[K]V, len(c.items))
 	now := time.Now()
 	for k, item := range c.items {
 		if !checkExpired || c.has(k, &now) {
-			items[k] = item.value
+			value, err := c.deserializeValue(k, item.value)
+			if err != nil {
+				// TODO: log error
+				continue
+			}
+			items[k] = value
 		}
 	}
 	return items
 }
 
 // Keys returns a slice of the keys in the cache.
-func (c *SimpleCache) Keys(checkExpired bool) []interface{} {
+func (c *SimpleCache[K, V, S]) Keys(checkExpired bool) []K {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
-	keys := make([]interface{}, 0, len(c.items))
+	keys := make([]K, 0, len(c.items))
 	now := time.Now()
 	for k := range c.items {
 		if !checkExpired || c.has(k, &now) {
@@ -258,7 +263,7 @@ func (c *SimpleCache) Keys(checkExpired bool) []interface{} {
 }
 
 // Len returns the number of items in the cache.
-func (c *SimpleCache) Len(checkExpired bool) int {
+func (c *SimpleCache[K, V, S]) Len(checkExpired bool) int {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
 	if !checkExpired {
@@ -275,27 +280,25 @@ func (c *SimpleCache) Len(checkExpired bool) int {
 }
 
 // Completely clear the cache
-func (c *SimpleCache) Purge() {
+func (c *SimpleCache[K, V, S]) Purge() {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 
-	if c.purgeVisitorFunc != nil {
-		for key, item := range c.items {
-			c.purgeVisitorFunc(key, item.value)
-		}
+	for key, item := range c.items {
+		err := c.purgeValue(key, item.value)
+		_ = err // TODO: log error
 	}
-
 	c.init()
 }
 
-type simpleItem struct {
+type simpleItem[V any] struct {
 	clock      Clock
-	value      interface{}
+	value      V
 	expiration *time.Time
 }
 
 // IsExpired returns boolean value whether this item is expired or not.
-func (si *simpleItem) IsExpired(now *time.Time) bool {
+func (si *simpleItem[V]) IsExpired(now *time.Time) bool {
 	if si.expiration == nil {
 		return false
 	}
