@@ -6,11 +6,11 @@ import (
 )
 
 func main() {
-	gc := gcache.New(10).
+	gc := gcache.New[string, string](10).
 		LFU().
-		LoaderFunc(func(key interface{}) (interface{}, error) {
-		return fmt.Sprintf("%v-value", key), nil
-	}).
+		LoaderFunc(func(key string) (string, error) {
+			return fmt.Sprintf("%v-value", key), nil
+		}).
 		Build()
 
 	v, err := gc.Get("key")
