@@ -30,18 +30,18 @@ func TestStats(t *testing.T) {
 	}
 }
 
-func getter(key interface{}) (interface{}, error) {
+func getter(key int) (int, error) {
 	return key, nil
 }
 
 func TestCacheStats(t *testing.T) {
 	var cases = []struct {
-		builder func() Cache
+		builder func() Cache[int, int]
 		rate    float64
 	}{
 		{
-			builder: func() Cache {
-				cc := New(32).Simple().Build()
+			builder: func() Cache[int, int] {
+				cc := New[int, int](32).Simple().Build()
 				cc.Set(0, 0)
 				cc.Get(0)
 				cc.Get(1)
@@ -50,8 +50,8 @@ func TestCacheStats(t *testing.T) {
 			rate: 0.5,
 		},
 		{
-			builder: func() Cache {
-				cc := New(32).LRU().Build()
+			builder: func() Cache[int, int] {
+				cc := New[int, int](32).LRU().Build()
 				cc.Set(0, 0)
 				cc.Get(0)
 				cc.Get(1)
@@ -60,8 +60,8 @@ func TestCacheStats(t *testing.T) {
 			rate: 0.5,
 		},
 		{
-			builder: func() Cache {
-				cc := New(32).LFU().Build()
+			builder: func() Cache[int, int] {
+				cc := New[int, int](32).LFU().Build()
 				cc.Set(0, 0)
 				cc.Get(0)
 				cc.Get(1)
@@ -70,8 +70,8 @@ func TestCacheStats(t *testing.T) {
 			rate: 0.5,
 		},
 		{
-			builder: func() Cache {
-				cc := New(32).ARC().Build()
+			builder: func() Cache[int, int] {
+				cc := New[int, int](32).ARC().Build()
 				cc.Set(0, 0)
 				cc.Get(0)
 				cc.Get(1)
@@ -80,8 +80,8 @@ func TestCacheStats(t *testing.T) {
 			rate: 0.5,
 		},
 		{
-			builder: func() Cache {
-				cc := New(32).
+			builder: func() Cache[int, int] {
+				cc := New[int, int](32).
 					Simple().
 					LoaderFunc(getter).
 					Build()
@@ -93,8 +93,8 @@ func TestCacheStats(t *testing.T) {
 			rate: 0.5,
 		},
 		{
-			builder: func() Cache {
-				cc := New(32).
+			builder: func() Cache[int, int] {
+				cc := New[int, int](32).
 					LRU().
 					LoaderFunc(getter).
 					Build()
@@ -106,8 +106,8 @@ func TestCacheStats(t *testing.T) {
 			rate: 0.5,
 		},
 		{
-			builder: func() Cache {
-				cc := New(32).
+			builder: func() Cache[int, int] {
+				cc := New[int, int](32).
 					LFU().
 					LoaderFunc(getter).
 					Build()
@@ -119,8 +119,8 @@ func TestCacheStats(t *testing.T) {
 			rate: 0.5,
 		},
 		{
-			builder: func() Cache {
-				cc := New(32).
+			builder: func() Cache[int, int] {
+				cc := New[int, int](32).
 					ARC().
 					LoaderFunc(getter).
 					Build()
